// Stocker/ProjetoLDS/Stocker/Frontend/StockerWebUI/src/app/enums/types.enum.ts
// Enum que define os tipos de produtos. Cada valor do enum é representado por uma string descritiva.
export enum Types {
  Frutas_Verduras = 'Frutas_Verduras',
  Panificacao_Confeitaria = 'Panificacao_Confeitaria',
  Laticinios = 'Laticinios',
  Carne_Peixe = 'Carne_Peixe',
  Ingredientes_Temperos = 'Ingredientes_Temperos',
  Congelados = 'Congelados',
  Cereais_Graos = 'Cereais_Graos',
  Lanches_Doces = 'Lanches_Doces',
  Bebidas = 'Bebidas',
  Casa = 'Casa',
  Higiene_Saude = 'Higiene_Saude',
  Animais = 'Animais',
  Artesanato_Jardim = 'Artesanato_Jardim',
  Outro = 'Outro',
}